+++ src/enums/CONFIGLEVEL.ts
@@ -1,0 +1,8 @@
+
+export enum CONFIGLEVEL {
+    DEFAULT = "default",
+    ENVIRONMENT = "environment",
+    BACKEND = "backend",
+    USER = "user",
+    DYNAMIC = "dynamic"
+}