+++ src/tests/backend-update.test.ts
@@ -1,11 +1,12 @@
 import { beforeEach, describe, expect, test, vi } from "vitest";
-import { IConfigurationObject } from "../interfaces/IConfigurationObject";
+import { IConfig } from "../interfaces/IConfigurationObject";
 import { Configuration } from "../Configuration.class";
 import { ConfigurationError } from "../ConfigurationError";
+import { CONFIGLEVEL } from "../enums/CONFIGLEVEL";
 
-interface TestConfig extends IConfigurationObject {
-    settingA: string;
-    settingB: number;
+interface TestConfig extends IConfig {
+    foo: string;
+    bar: number;
 }
 
 describe("Backend update process", () => {
@@ -14,10 +15,10 @@ describe("Backend update process", () => {
     });
 
     test("Start backend auto-update process successfully", async () => {
-        const update: Partial<TestConfig> = { settingA: "updatedA" };
+        const update: Partial<TestConfig> = { foo: "updatedA" };
 
         const conf = new Configuration<TestConfig>(
-            { settingA: "defaultA", settingB: 10 },
+            { foo: "defaultA", bar: 10 },
             {
                 backendUpdateFn: <TestConfig>(): Promise<Partial<TestConfig>> => {
                     return Promise.resolve(update as Partial<TestConfig>);
@@ -27,21 +28,21 @@ describe("Backend update process", () => {
             }
         );
 
-        expect(conf.getValue("settingA")).toBe("defaultA");
+        expect(conf.getValue("foo")).toBe("defaultA");
 
         // Wait for some time to allow the update to occur
         await new Promise((resolve) => setTimeout(resolve, 50));
 
-        expect(conf.getValue("settingA")).toBe("updatedA");
+        expect(conf.getValue("foo")).toBe("updatedA");
 
         conf.stopBackendAutoUpdate();
     });
 
     test("Backend updates must trigger subscribers", async () => {
-        const update: Partial<TestConfig> = { settingB: 42 };
+        const update: Partial<TestConfig> = { bar: 42 };
 
         const conf = new Configuration<TestConfig>(
-            { settingA: "defaultA", settingB: 10 },
+            { foo: "defaultA", bar: 10 },
             {
                 backendUpdateFn: <TestConfig>(): Promise<Partial<TestConfig>> => {
                     return Promise.resolve(update as Partial<TestConfig>);
@@ -53,9 +54,9 @@ describe("Backend update process", () => {
 
         const cb = vi.fn((changedKeys: Partial<TestConfig>) => {});
 
-        conf.subscribe(["settingB"], cb);
+        conf.subscribe(["bar"], cb);
 
-        expect(conf.getValue("settingB")).toBe(10);
+        expect(conf.getValue("bar")).toBe(10);
 
         conf.startBackendAutoUpdate();
 
@@ -64,22 +65,22 @@ describe("Backend update process", () => {
         // Wait for some time to allow the update to occur
         await new Promise((resolve) => setTimeout(resolve, 150));
 
-        expect(conf.getValue("settingB")).toBe(42);
-        expect(cb).toHaveBeenCalledWith({ settingB: 42 });
+        expect(conf.getValue("bar")).toBe(42);
+        expect(cb).toHaveBeenCalledWith({ bar: { value: 42, level: CONFIGLEVEL.BACKEND, readonly: false } });
         expect(cb).toHaveBeenCalledTimes(1);
 
         conf.stopBackendAutoUpdate();
     });
 
-    test("Backend updates must trigger until stopped", async () => {
-        const update: Partial<TestConfig> = { settingB: 42 };
+    test.only("Backend updates must trigger until stopped", async () => {
+        const update: Partial<TestConfig> = { bar: 42 };
 
         const conf = new Configuration<TestConfig>(
-            { settingA: "defaultA", settingB: 10 },
+            { foo: "defaultA", bar: 10 },
             {
                 backendUpdateFn: (cc: TestConfig): Promise<Partial<TestConfig>> => {
-                    const curB = cc.settingB;
-                    const newConf: Partial<TestConfig> = { settingB: curB + 1 };
+                    const newConf = Configuration.helperConvertToValueObject(cc);
+                    newConf.bar = (newConf.bar as number) + 1;
                     return Promise.resolve(newConf);
                 },
                 backendUpdateIntervalMs: 50,
@@ -89,9 +90,9 @@ describe("Backend update process", () => {
 
         const cb = vi.fn((changedKeys: Partial<TestConfig>) => {});
 
-        conf.subscribe(["settingB"], cb);
+        conf.subscribe(["bar"], cb);
 
-        expect(conf.getValue("settingB")).toBe(10);
+        expect(conf.getValue("bar")).toBe(10);
 
         conf.startBackendAutoUpdate();
 
@@ -103,13 +104,13 @@ describe("Backend update process", () => {
         conf.stopBackendAutoUpdate();
 
         expect(cb).toHaveBeenCalledTimes(4);
-        expect(conf.getValue("settingB")).toBe(14);
+        expect(conf.getValue("bar")).toBe(14);
 
         // Wait to ensure no more updates occur
         await new Promise((resolve) => setTimeout(resolve, 100));
 
         expect(cb).toHaveBeenCalledTimes(4);
-        expect(conf.getValue("settingB")).toBe(14);
+        expect(conf.getValue("bar")).toBe(14);
 
         // Restart the update process
         conf.startBackendAutoUpdate();
@@ -117,14 +118,14 @@ describe("Backend update process", () => {
         await new Promise((resolve) => setTimeout(resolve, 60));
 
         expect(cb).toHaveBeenCalledTimes(6);
-        expect(conf.getValue("settingB")).toBe(16);
+        expect(conf.getValue("bar")).toBe(16);
 
         conf.stopBackendAutoUpdate();
     }, 1000);
 
     test("If backend update function fails, error is caught and no update occurs", async () => {
         const conf = new Configuration<TestConfig>(
-            { settingA: "defaultA", settingB: 10 },
+            { foo: "defaultA", bar: 10 },
             {
                 backendUpdateFn: (): Promise<Partial<TestConfig>> => {
                     throw new SyntaxError("Simulated backend failure");
@@ -134,7 +135,7 @@ describe("Backend update process", () => {
             }
         );
 
-        expect(conf.getValue("settingB")).toBe(10);
+        expect(conf.getValue("bar")).toBe(10);
 
         await expect(conf.startBackendAutoUpdate()).rejects.toThrow(ConfigurationError);